Refactoring auth code into middleware
